requirements.txt
numpy
imageio
imageio-ffmpeg
moviepy
pygame-ce
pygbag
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
opencv-python
opencv-python-headless
Pillow==9.5.0
proglog==0.1.10
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.2

